+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## Unreleased
+## [0.1.4] - 2023-10-10
 
 ### Added
 
@@ -52,6 +52,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.1.1]: https://github.com/dogmatiq/primo/releases/tag/v0.1.1
 [0.1.2]: https://github.com/dogmatiq/primo/releases/tag/v0.1.2
 [0.1.3]: https://github.com/dogmatiq/primo/releases/tag/v0.1.3
+[0.1.4]: https://github.com/dogmatiq/primo/releases/tag/v0.1.4
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD